refactor(backend): move tmdb utils to module
