+++ apps/backend/src/providers/tmdb.rs
@@ -15,11 +15,7 @@ use crate::{
     migrator::{MetadataImageLot, MetadataLot, MetadataSource},
     models::{MovieSpecifics, ShowEpisode, ShowSeason, ShowSpecifics},
     traits::MediaProvider,
-    utils::{
-        convert_date_to_year, convert_string_to_date,
-        tmdb::{self, save_all_images, TmdbCredit},
-        NamedObject,
-    },
+    utils::{convert_date_to_year, convert_string_to_date, NamedObject},
 };
 
 #[derive(Debug, Clone)]
@@ -30,7 +26,7 @@ pub struct MovieTmdbService {
 
 impl MovieTmdbService {
     pub async fn new(config: &MoviesTmdbConfig) -> Self {
-        let (client, image_url) = tmdb::get_client_config(&config.url, &config.access_token).await;
+        let (client, image_url) = utils::get_client_config(&config.url, &config.access_token).await;
         Self { client, image_url }
     }
 }
@@ -57,8 +53,8 @@ impl MediaProvider for MovieTmdbService {
         let data: TmdbMovie = rsp.body_json().await.map_err(|e| anyhow!(e))?;
         #[derive(Debug, Serialize, Deserialize, Clone)]
         struct TmdbCreditsResponse {
-            cast: Vec<TmdbCredit>,
-            crew: Vec<TmdbCredit>,
+            cast: Vec<utils::TmdbCredit>,
+            crew: Vec<utils::TmdbCredit>,
         }
         let mut rsp = self
             .client
@@ -102,7 +98,7 @@ impl MediaProvider for MovieTmdbService {
         if let Some(u) = data.backdrop_path {
             image_ids.push(u);
         }
-        save_all_images(&self.client, "movie", identifier, &mut image_ids).await?;
+        utils::save_all_images(&self.client, "movie", identifier, &mut image_ids).await?;
 
         Ok(MediaDetails {
             identifier: data.id.to_string(),
@@ -199,7 +195,7 @@ pub struct ShowTmdbService {
 
 impl ShowTmdbService {
     pub async fn new(config: &ShowsTmdbConfig) -> Self {
-        let (client, image_url) = tmdb::get_client_config(&config.url, &config.access_token).await;
+        let (client, image_url) = utils::get_client_config(&config.url, &config.access_token).await;
         Self { client, image_url }
     }
 }
@@ -232,7 +228,7 @@ impl MediaProvider for ShowTmdbService {
         if let Some(u) = data.backdrop_path {
             image_ids.push(u);
         }
-        save_all_images(&self.client, "tv", identifier, &mut image_ids).await?;
+        utils::save_all_images(&self.client, "tv", identifier, &mut image_ids).await?;
 
         #[derive(Debug, Serialize, Deserialize, Clone)]
         struct TmdbEpisode {
@@ -243,8 +239,8 @@ impl MediaProvider for ShowTmdbService {
             overview: Option<String>,
             air_date: Option<String>,
             runtime: Option<i32>,
-            guest_stars: Vec<TmdbCredit>,
-            crew: Vec<TmdbCredit>,
+            guest_stars: Vec<utils::TmdbCredit>,
+            crew: Vec<utils::TmdbCredit>,
         }
         #[derive(Debug, Serialize, Deserialize, Clone)]
         struct TmdbSeason {
@@ -441,5 +437,92 @@ impl MediaProvider for ShowTmdbService {
 impl ShowTmdbService {
     fn get_cover_image_url(&self, c: &str) -> String {
         format!("{}{}{}", self.image_url, "original", c)
+    }
+}
+
+mod utils {
+    use std::{env, fs};
+
+    use surf::{
+        http::headers::{AUTHORIZATION, USER_AGENT},
+        Config, Url,
+    };
+
+    use crate::{
+        graphql::{AUTHOR, PROJECT_NAME},
+        utils::read_file_to_json,
+    };
+
+    use super::*;
+
+    #[derive(Debug, Serialize, Deserialize, Clone)]
+    pub struct TmdbCredit {
+        pub name: Option<String>,
+        pub known_for_department: Option<String>,
+        pub profile_path: Option<String>,
+    }
+
+    #[derive(Debug, Serialize, Deserialize, Clone)]
+    pub struct TmdbImage {
+        pub file_path: String,
+    }
+
+    #[derive(Debug, Serialize, Deserialize, Clone)]
+    pub struct TmdbImagesResponse {
+        pub backdrops: Option<Vec<utils::TmdbImage>>,
+        pub posters: Option<Vec<utils::TmdbImage>>,
+    }
+
+    pub async fn get_client_config(url: &str, access_token: &str) -> (Client, String) {
+        let path = env::temp_dir().join("tmdb-config.json");
+        let client: Client = Config::new()
+            .add_header(USER_AGENT, format!("{}/{}", AUTHOR, PROJECT_NAME))
+            .unwrap()
+            .add_header(AUTHORIZATION, format!("Bearer {access_token}"))
+            .unwrap()
+            .set_base_url(Url::parse(url).unwrap())
+            .try_into()
+            .unwrap();
+        #[derive(Debug, Serialize, Deserialize, Clone)]
+        struct TmdbImageConfiguration {
+            secure_base_url: String,
+        }
+        #[derive(Debug, Serialize, Deserialize, Clone)]
+        struct TmdbConfiguration {
+            images: TmdbImageConfiguration,
+        }
+        let image_url = if let Some(details) = read_file_to_json::<TmdbConfiguration>(&path) {
+            details.images.secure_base_url
+        } else {
+            let mut rsp = client.get("configuration").await.unwrap();
+            let data: TmdbConfiguration = rsp.body_json().await.unwrap();
+            fs::write(path, serde_json::to_string(&data).unwrap()).ok();
+            data.images.secure_base_url
+        };
+        (client, image_url)
+    }
+
+    pub async fn save_all_images(
+        client: &Client,
+        typ: &str,
+        identifier: &str,
+        images: &mut Vec<String>,
+    ) -> Result<()> {
+        let mut rsp = client
+            .get(format!("{}/{}/images", typ, identifier))
+            .await
+            .map_err(|e| anyhow!(e))?;
+        let new_images: TmdbImagesResponse = rsp.body_json().await.map_err(|e| anyhow!(e))?;
+        if let Some(imgs) = new_images.posters {
+            for image in imgs {
+                images.push(image.file_path);
+            }
+        }
+        if let Some(imgs) = new_images.backdrops {
+            for image in imgs {
+                images.push(image.file_path);
+            }
+        }
+        Ok(())
     }
 }
